change layout of image, use image2DArray instead
This will help simplify the logical of NCHW,
image will pack as (W, H * N, C4) where C4 is the arraylayers.
For input N will be 1 for most case, so the convertor is quite
straightfull. And memory is continuous for all channel, since each layer
will have a row-continuous and height * N images

Signed-off-by: junka <wan.junjie@foxmail.com>